sensor-server-viewer/requirements.txt
pyzmq==26.2.0
nicegui==2.5.0
rel==0.4.9.19
websocket-client==1.8.0
requests==2.32.3
nicegui-highcharts==2.0.2
